- forgot to set the Enter callback in the detail fragment. also forgot to set the mArticleDetailFragment, in order to use set the ImageView as the shared element in the callback. - also forgot to implement the callback in the listactivity. - removed the adding to the ARTICLES constant ArrayList, as I was already creating it when parsing the UpdaterService. This was causing the data to be innacurrate and thus breaking the return shared element transitions
This callback checks if the user is returning by checking if the mTmpReenterState exists. If it exists then it means we have a startingPosition and currentPosition and that we’re returning to the activity. We check the starting vs current positions to determine whether or not to create a newTransitionName. We then map the transitionnames.